Replaced virtual logic with extendable
Since a mailer is wrapped by the IMAP protocol,
there is no need to rewrite the same logic for
every known mailer.

However, it is possible one of them might use
the IMAP protocol differently. For that reason,
it is still possible to override methods of the
extended Mailer class.